+++ py/desispec/scripts/fluxcalibration.py
@@ -37,6 +37,8 @@ def parse(options=None):
                         help = 'apply a reduced chi2 cut for the selection of stars')
     parser.add_argument('--chi2cut-nsig', type = float, default = 3., required=False,
                         help = 'discard n-sigma outliers from the reduced chi2 of the standard star fit')
+    parser.add_argument('--delta-color-cut', type = float, default = 0.1, required=False,
+                        help = 'discard model stars with different broad-band color from imaging')
     parser.add_argument('--outfile', type = str, default = None, required=True,
                         help = 'path of DESI flux calbration fits file')
     parser.add_argument('--qafile', type=str, default=None, required=False,
@@ -92,6 +94,17 @@ def main(args) :
             model_fibers=model_fibers[ok]
             model_metadata=model_metadata[:][ok]
     
+    if args.delta_color_cut > 0 :
+        ok = np.where(np.abs(model_metadata["MODEL_G-R"]-model_metadata["DATA_G-R"])<args.delta_color_cut)[0]
+        nstars=model_flux.shape[0]
+        nbad=nstars-ok.size
+        if nbad>0 :
+            log.warning("discarding %d star(s) out of %d because |delta_color|>%f"%(nbad,nstars,args.delta_color_cut))
+            model_flux=model_flux[ok]
+            model_fibers=model_fibers[ok]
+            model_metadata=model_metadata[:][ok]
+    
+
     # automatically reject stars that ar chi2 outliers
     if args.chi2cut_nsig > 0 :
         mchi2=np.median(model_metadata["CHI2DOF"])